moving rag UI to a central spot
