+++ config/routes.rb
@@ -54,7 +54,7 @@ Discourse::Application.routes.draw do
     ) { post :test, on: :collection }
 
     post "/ai-personas/:id/create-user", to: "discourse_ai/admin/ai_personas#create_user"
-    post "/ai-personas/files/upload", to: "discourse_ai/admin/ai_personas#upload_file"
+    post "/rag-document-fragments/files/upload", to: "discourse_ai/admin/rag_document_fragments#upload_file"
     put "/ai-personas/:id/files/remove", to: "discourse_ai/admin/ai_personas#remove_file"
     get "/ai-personas/:id/files/status", to: "discourse_ai/admin/ai_personas#indexing_status_check"
